Added userRole, test cases for controller and Repositories

+++ src/main/java/com/studentManagementSystem/services/StudentService.java
@@ -1,6 +1,7 @@
 package com.studentManagementSystem.services;
 
 import com.studentManagementSystem.entities.Student;
+import com.studentManagementSystem.constants.AppConstants;
 import com.studentManagementSystem.repositories.StudentRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -23,28 +24,27 @@ public class StudentService {
      * @return a string confirming successful submission or not
      *
      **/
-    public String submitStudentDetails(Student student)
-    {
+    public String submitStudentDetails(String role, Student student) throws Exception {
+        if(role.equalsIgnoreCase(AppConstants.ADMIN)) {
+            studentList = studentRepository.findByUserId(student.getUserId());
+            try {
+                if (isNull(student)) {
+                    return "Enter student details";
+                } else if (!studentList.isEmpty()) {
+                    return "Student already exists";
+                } else if (student.getName() == null || student.getSurname() == null || student.getCellNumber() == null) {
+                    return "There's a missing mandatory field";
+                } else {
+                    studentRepository.save(student);
 
-        studentList = studentRepository.findByStudentId(student.getStudentId());
-       try {
-           if (isNull(student)) {
-               return "Enter student details";
-           }
-           else if(!studentList.isEmpty()){
-               return "Student already exists";
-           }
-           else
-               if (student.getName() == null || student.getSurname() == null || student.getCellNumber() == null) {
-               return "There's a missing mandatory field";
-           } else {
-           studentRepository.save(student);
-
-           }
-       }catch (Exception e){
-           e.getMessage();
-       }
-           return "Successfully saved Student";
+                }
+            } catch (Exception e) {
+                e.getMessage();
+            }
+            return "Successfully saved Student";
+        }else {
+            throw new Exception("You do not have the rights to add student");
+        }
 
     }
 
@@ -54,16 +54,20 @@ public class StudentService {
      * @return a List of students
      *
      **/
-    public List<Student> retrieveAllStudents()
-    {
-        try {
-            studentList = studentRepository.findAll();
-        }catch (ArrayIndexOutOfBoundsException E){
-            E.getMessage();
-        }catch (Exception Ex){
-            Ex.getMessage();
+    public List<Student> retrieveAllStudents(String role) throws Exception {
+        if (role.equalsIgnoreCase(AppConstants.ADMIN)) {
+            try {
+                studentList = studentRepository.findAll();
+            } catch (ArrayIndexOutOfBoundsException E) {
+                E.getMessage();
+            } catch (Exception Ex) {
+                Ex.getMessage();
+            }
+            return studentList;
         }
-        return studentList;
+        else {
+            throw new Exception("You do not have the rights to retrieve all students");
+        }
     }
 
     /**
@@ -73,16 +77,18 @@ public class StudentService {
      * @return a string confirming successful update or not
      *
      **/
-    public String updateStudentDetails(Long studentId, Student student)
-    {
-        studentList = studentRepository.findByStudentId(studentId);
-        if (studentList.isEmpty()) {
-            return "Student is not registered";
+    public String updateStudentDetails(String role, Long studentId, Student student) throws Exception {
+        if (role.equalsIgnoreCase(AppConstants.ADMIN) || role.equalsIgnoreCase(AppConstants.STUDENT)) {
+            studentList = studentRepository.findByUserId(studentId);
+            if (studentList.isEmpty()) {
+                return "Student is not registered";
+            } else {
+                studentRepository.save(student);
+                return "Successfully updated Student";
+            }
         }
-        else{
-            studentRepository.delete(studentList.get(0));
-            studentRepository.save(student);
-            return "Successfully updated Student";
+        else {
+            throw new Exception("You do not have the rights to update student details");
         }
 
     }
@@ -94,15 +100,18 @@ public class StudentService {
      * @return a string confirming successful delete
      *
      **/
-    public String deleteStudent(Long studentId)
-    {
-        studentList = studentRepository.findByStudentId(studentId);
-        if (studentList.isEmpty()) {
-            return "Student is not registered";
+    public String deleteStudent(String role, Long studentId) throws Exception {
+        if (role.equalsIgnoreCase(AppConstants.ADMIN) || role.equalsIgnoreCase(AppConstants.STUDENT)) {
+            studentList = studentRepository.findByUserId(studentId);
+            if (studentList.isEmpty()) {
+                return "Student is not registered";
+            } else {
+                studentRepository.delete(studentList.get(0));
+                return "Successfully de-registered Student";
+            }
         }
-        else{
-            studentRepository.delete(studentList.get(0));
-            return "Successfully de-registered Student";
+        else {
+            throw new Exception("You do not have the rights to de-register student");
         }
 
     }
